% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Cost_Complexity_Pruning.R
\name{best_node_cost}
\alias{best_node_cost}
\title{best_node_cost}
\usage{
best_node_cost(node, tree)
}
\arguments{
\item{node}{}

\item{tree}{}
}
\description{
function that gives the minimum change in tree cost ==> gives alpha = min g(t), for all t in T
}
\details{
WARNINGS: this function was modified on the 4th december 2017 to be adapted to non-binary trees.
}
