% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions_RF_evo_Modif_pierre.R
\name{bsamples}
\alias{bsamples}
\title{bsamples}
\usage{
bsamples(ntree, data, sampsize, replace)
}
\arguments{
\item{ntree}{threes number}

\item{data}{data sample}

\item{sampsize}{integer}

\item{replace}{boolean}
}
\value{
a matrix which with sampsize lines and B columns which
contains the indices of the observations belonging each boostrap sample.
}
\description{
Bootstrap samples
}
\details{
Function which takes as inputs two integers "B" and "sampsize",
a sample of data "data" and a boolean "replace".
The function draws randomly B bootstrap samples of size "sampsize"
(the size of the sample of data)
}
