% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carts}
\alias{carts}
\title{Data to test function predict_cartgv}
\format{A list of data frames}
\usage{
carts
}
\description{
A dataset generated to check that fucntion predict_cartgv works.
}
\keyword{datasets}
