% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Cost_Complexity_Pruning.R
\name{conditional_error}
\alias{conditional_error}
\title{conditional_error}
\usage{
conditional_error(node, tree)
}
\arguments{
\item{node}{}

\item{tree}{}
}
\description{
function that calculate the missclassification error assuming that we are in the node
}
