% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Cost_Complexity_Pruning.R
\name{cost_complexity_binary}
\alias{cost_complexity_binary}
\title{cost_complexity_binary}
\usage{
cost_complexity_binary(tree)
}
\arguments{
\item{tree}{}
}
\description{
cost_complexity_binary
}
