% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Cost_Complexity_Pruning.R
\name{cost_error}
\alias{cost_error}
\title{cost_error}
\usage{
cost_error(node, tree, alpha)
}
\arguments{
\item{node}{}

\item{tree}{}

\item{alpha}{}
}
\description{
function that calculate the misclassification error of the subtree stremming from node penalized by alpha
}
