% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Data to test function rfgv}
\format{A data frame with 1000 rows and 27 variables:

\describe{
\item{Y}{0 or 1}
\item{VxxG1}{vector values of group 1}
\item{VyyG2}{vector values of group 2}
}}
\usage{
data
}
\description{
A dataset generated to check that fucntion rfgv works.
}
\keyword{datasets}
