% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Cost_Complexity_Pruning.R
\name{descendant}
\alias{descendant}
\title{descendant}
\usage{
descendant(node, tree)
}
\arguments{
\item{node}{}

\item{tree}{}
}
\description{
function that find the direct descendants of a node
}
