% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_RFGV.R
\name{entropy}
\alias{entropy}
\title{entropy}
\usage{
entropy(p)
}
\arguments{
\item{p}{a vector containit the probabilities '\code{P[Y=0]}' and '\code{P[Y=1]}'}
}
\value{
the value of the entropy
}
\description{
Calculate the value of the entropy for binary classification (i.e. Y= 0 or 1)
}
