% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Cost_Complexity_Pruning.R
\name{extract_subtrees}
\alias{extract_subtrees}
\title{extratct_subtrees}
\usage{
extract_subtrees(tree, cp)
}
\arguments{
\item{tree}{a tree. A data frame returned by the function \code{CARTGV} and which summarizes the resulted CARTGV tree.}

\item{cp}{a data frame that summarizes the complexity as explained in the example in Breiman (chap 10, 1984). It is the object \code{cp} returned by the function \code{cost_complexity}.}
}
\value{
a list containing the sequence of optimal subtrees.
}
\description{
function that extract the sequence of nested trees
}
