% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Cost_Complexity_Pruning.R
\name{get_node_leaves}
\alias{get_node_leaves}
\title{get_node_leaves}
\usage{
get_node_leaves(node, tree)
}
\arguments{
\item{node}{}

\item{tree}{}
}
\description{
function that gives the leaves of the subtree stremming from node
}
