% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Cost_Complexity_Pruning.R
\name{get_nodes}
\alias{get_nodes}
\title{get_nodes}
\usage{
get_nodes(node, tree)
}
\arguments{
\item{node}{the number of the node in the tree \code{tree}.}

\item{tree}{a tree. A data frame returned by the function \code{CARTGV} and which summarizes the resulted CARTGV tree.}
}
\value{
a vector indicating the nodes of the subtree stremming from \code{node}.
}
\description{
function that gives the nodes of the subtree stremming from node
}
