% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_RFGV.R
\name{gini}
\alias{gini}
\title{gini}
\usage{
gini(p)
}
\arguments{
\item{p}{p a vector containit the probabilities '\code{P[Y=0]}' and '\code{P[Y=1]}'}
}
\value{
the value of the Gini index
}
\description{
Calculate the value of the Gini index for binary classification (i.e. Y= 0 or 1)
}
