% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPhoneme_test.R, R/group.R
\docType{data}
\name{group}
\alias{group}
\title{Dataframe contains in data_phoneme}
\format{data frame}
\usage{
group

group
}
\description{
A dataset generated to check that fucntion rfgv works.
Please load with rfgv_dataset
}
\details{
26 values
}
\keyword{datasets}
