% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions_RF_evo_Modif_pierre.R
\name{group.selection}
\alias{group.selection}
\title{group.selection}
\usage{
group.selection(group, mtry = sqrt(unique(group[!is.na(group)])))
}
\arguments{
\item{group}{group that contain each variable}

\item{mtry}{m try}
}
\value{
groups
}
\description{
Function which takes inputs the vector "group" which indicates the group that contain each
variable and a number "mtry" (m<=length(group)).
The function selects randomly mtry groups among the length(group) groups and returns a
vector with the indices of the mtry selected group.
note that the default value for m is sqrt(length(group))
}
