% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions_CARTGVRF_sampvar1.R
\name{grpimpperm}
\alias{grpimpperm}
\title{Title}
\usage{
grpimpperm(num_group, data, oobsamples, group, tree, impurityacc)
}
\arguments{
\item{num_group}{}

\item{data}{}

\item{oobsamples}{}

\item{group}{}

\item{tree}{}

\item{impurityacc}{}
}
\description{
Title
}
