% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_variant_CARTGV_for_RFGV.R
\name{grpimpperm.rf}
\alias{grpimpperm.rf}
\title{grpimpperm.rf}
\usage{
grpimpperm.rf(num_group, data, oobsamples, group, tree, impurityacc)
}
\arguments{
\item{num_group}{index of the considered group}

\item{data}{a data frame containing all observation included in the training data set.}

\item{oobsamples}{a vector containing the indices of the observations that are not included in the sample used to build the modified CARTGV tree.}

\item{group}{a vector with the group number of each variable. (WARNING : if there are "\code{p}" goups, the groups must be numbers
from "\code{1}" to "\code{p}" in increasing order. The group label of the response variable is missing (i.e. NA)).}

\item{tree}{the output of the function \code{cartgv.rf}.}

\item{impurityacc}{the accurancy value of the modified CARTGV tree.}
}
\value{
DecreaseAcc the decrease in accurance from permuting the values of the considered group.
}
\description{
Compute the permutation importance of a group from a modified CARTGV tree. Used only by the function \code{rfgv} when grp.importance=TRUE.
}
