% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions_CARTGVRF_sampvar1.R
\name{impurity.cartgv}
\alias{impurity.cartgv}
\title{Title
Elagage :  calcul de l'impuret'e a' partir d'un 'echantillon independant et d'une s'equence d'arbres emboites
WARNING : fonction corrigée le 19 janvier2018 car il y avait un problème d'ordre pour les noeuds dans les apply
La fonction prend en entrée un échantillon test et une séquence d'arbres emboités
Pour chaque arbre, on prédit la classe de chaque observation de l'ensemble test;
puis à partir de ces résultats, on calcul l'impureté de chaque arbre (gini, l'entropie et le taux de mal-classés sont calculés)
method : la nature des coupures pour l'arbre CC binaire ("bin" si tree construit avec Tree_CART_bin())
ou ("multi" si tree construit avec Tree_CART())}
\usage{
impurity.cartgv(validation, tree_seq, tree)
}
\arguments{
\item{validation}{data.frame contenant les mêmes variables que le jeux de données utilisés pour construire l'objet tree}

\item{tree_seq}{objet retourné par la fonction Tree_CART_bin() ou Tree_CART()}

\item{tree}{}
}
\value{
List :
impurete : une matrice contenant les differentes valeurs d'impuret'e pour chaque sous-arbres
pred = liste des predictions pour chaque sous arbres
summary_noeuds : liste contenant pour chaque sous-arbre des infos sur ses noeuds:
nom_noeuds = noms du noeuds
N = nb obs dans le neoud
N\link{Y=1} = nb d'obs dans le noeuds appartenant a' la class "Y=1"
P\link{Y=1} = proba pour une obs du noeuds d'appartenanir  a' la class "Y=1"
P\link{Y=0} = proba pour une obs du noeuds d'appartenanir  a' la class "Y=0"
P\link{hat.Y!=Y} = tx de mal class'es dans le noeud
}
\description{
Title
Elagage :  calcul de l'impuret'e a' partir d'un 'echantillon independant et d'une s'equence d'arbres emboites
WARNING : fonction corrigée le 19 janvier2018 car il y avait un problème d'ordre pour les noeuds dans les apply
La fonction prend en entrée un échantillon test et une séquence d'arbres emboités
Pour chaque arbre, on prédit la classe de chaque observation de l'ensemble test;
puis à partir de ces résultats, on calcul l'impureté de chaque arbre (gini, l'entropie et le taux de mal-classés sont calculés)
method : la nature des coupures pour l'arbre CC binaire ("bin" si tree construit avec Tree_CART_bin())
ou ("multi" si tree construit avec Tree_CART())
}
