% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Cost_Complexity_Pruning.R
\name{local_error}
\alias{local_error}
\title{local_error}
\usage{
local_error(node, tree)
}
\arguments{
\item{node}{the number of the node in the tree \code{tree}.}

\item{tree}{a tree. A data frame returned by the function \code{CARTGV} and which summarizes the resulted CARTGV tree.}
}
\value{
a numeric indicating the missclassification error that the node contribute.
}
\description{
function that calculate the missclassification error that the node contribute
}
