% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Cost_Complexity_Pruning.R
\name{local_error_0}
\alias{local_error_0}
\title{local_error_0}
\usage{
local_error_0(node, tree)
}
\arguments{
\item{node}{}

\item{tree}{}
}
\description{
function that calculate the missclassification error that the node contribute
}
