% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{magasin}
\alias{magasin}
\title{Data for decision tree tutorial}
\format{A data frame with 78 rows and 5 variables:
\describe{
\item{Prix}{ Item prices}
\item{Flashitude}{  degree of beauty}
\item{Branchitude}{ degree of Hype}
\item{Qualite}{ degree of robustness}
\item{Achat}{ number of sells}
}}
\usage{
magasin
}
\description{
A dataset from Lise Vaudor blog post about decision tree
}
\keyword{datasets}
