% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Cost_Complexity_Pruning.R
\name{node_cost}
\alias{node_cost}
\title{node_cost}
\usage{
node_cost(node, tree)
}
\arguments{
\item{node}{the number of the node in the tree \code{tree}}

\item{tree}{a tree. A data frame returned by the function \code{CARTGV} and which summarizes the resulted CARTGV tree.}
}
\value{
a numeric indicating the change in tree cost (difference between the node error and the local node error). It corresponds to the value of the complexity parameter \code{alpha}.
}
\description{
function that gives the change in tree cost (difference between the node error and the local node error)  ==> gives alpha (function g(t) in the Breiman's book)
}
