% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Cost_Complexity_Pruning.R
\name{node_error}
\alias{node_error}
\title{node_error}
\usage{
node_error(node, tree)
}
\arguments{
\item{node}{the number of the node in the tree \code{tree}.}

\item{tree}{a tree. A data frame returned by the function \code{CARTGV} and which summarizes the resulted CARTGV tree.}
}
\value{
a numeric indicating the misclassification error of the subtree stremming from node.
}
\description{
Function that calculate the misclassification error of the subtree stremming from node
}
