% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Cost_Complexity_Pruning.R
\name{node_prob}
\alias{node_prob}
\title{node_prob}
\usage{
node_prob(node, tree)
}
\arguments{
\item{node}{the number of the node in the tree \code{tree}.}

\item{tree}{a tree. A data frame returned by the function \code{CARTGV} and which summarizes the resulted CARTGV tree.}
}
\value{
a numeric indicating the probability that an observation falls into a node.
}
\description{
function that calculate the probability of being at the node
}
