% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_performance.R
\name{opt.cut}
\alias{opt.cut}
\title{opt.cut}
\usage{
opt.cut(perf)
}
\arguments{
\item{perf}{an S4 object of class performance with elements to create the ROC curve  (y.values: True positive rate, x.values: False positive rate and alpha.value: the threshold)}
}
\value{
cut.ind : the index of the optimal cut
}
\description{
Find optimal cutoff (for the trade-off Sensitivity/specificity)
}
