% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions_performance.R
\name{performances}
\alias{performances}
\title{performances}
\usage{
performances(score, test)
}
\arguments{
\item{score}{score}

\item{test}{test}
}
\value{
resultats
}
\description{
performances
}
