% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_variant_CARTGV_for_RFGV.R
\name{perm}
\alias{perm}
\title{perm}
\usage{
perm(oobsamples, data, num_group, group)
}
\arguments{
\item{oobsamples}{a vector containing the indices of the observations that are not included in the sample used to build the modified CARTGV tree.}

\item{data}{a data frame containing all observation included in the training data set.}

\item{num_group}{index of the considered group}

\item{group}{a vector with the group number of each variable. (WARNING : if there are "\code{p}" goups, the groups must be numbers
from "\code{1}" to "\code{p}" in increasing order. The group label of the response variable is missing (i.e. NA)).}
}
\value{
data_perm a data frame obtained after permuting the values of the considered group for observation belonging to the out-of-bag sample.
}
\description{
Function that permutate the values of a group of variables.
Note that all the variables of the group are permuted together in order to keep the relationship between variables within the group.
}
