% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_RPART_tools.R
\name{pred_cart}
\alias{pred_cart}
\title{pred_cart}
\usage{
pred_cart(new_obs, tableau)
}
\arguments{
\item{new_obs}{a new observation. It must be a vector containing the values of the variables of the learning data set used to built the rpart object}

\item{tableau}{object returns by the function 'calcul_cart'}
}
\value{
a vector with elements (in this order): number of the predicted node in the splitting tree, index of the predicted node in the object returned
by the function 'pred_cart', the label of the predicted node, the number of observations with "\code{Y=1}"
in the predicted node, the number of observations with "\code{Y=0}" in the predicted node and the empirical
probability that in the predicted node an observation belongs to the label "\code{Y=1}"
}
\description{
Function used to predict the label ("0" or "1") of a new observation from a fitted CARTGV object.
}
