% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions_CARTGVRF_sampvar1.R
\name{predict.cartgv}
\alias{predict.cartgv}
\title{Title
la fonction fait appel a la fonction pred_cart()
IMPORTANT : i_noeuds (indice du noeud dans l'arbre de coupure, cad indice donne par rpart)
et noeuds (indice du noeud dans l'arbre cartgv) utilisees comme cles primaires pour
identifier de manière unique un noeuds}
\usage{
\method{predict}{cartgv}(new, tree, carts, coups)
}
\arguments{
\item{new}{}

\item{tree}{}

\item{carts}{}

\item{coups}{}
}
\value{
Valeurs retournees par la fonction :
la fonction renvoie une matrice donnant pour chaque individu de new (dans l'ordre) :
hat.Y :la classe predite
Y : la classe d'appartenance
noeuds : le numero du noeud contenant l'observation
score : le score le l'observation
}
\description{
Title
la fonction fait appel a la fonction pred_cart()
IMPORTANT : i_noeuds (indice du noeud dans l'arbre de coupure, cad indice donne par rpart)
et noeuds (indice du noeud dans l'arbre cartgv) utilisees comme cles primaires pour
identifier de manière unique un noeuds
}
