% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions_CARTGVRF_sampvar2.R
\name{predict.cartgv.rf}
\alias{predict.cartgv.rf}
\title{predict.cartgv}
\usage{
\method{predict}{cartgv.rf}(new, tree, tree_split)
}
\arguments{
\item{new}{}

\item{tree}{}

\item{tree_split}{}
}
\description{
la fonction fait appel à la fonction predict.cartgv()
IMPORTANT : i_noeuds (indice du noeud dans l'arbre de coupure,
càd indice donné par rpart) et noeuds (indice du noeud dans l'arbre cartgv)
utilisées comme clés primaires pour identifier de manière unique un noeuds
}
