% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions_CARTGVRF_sampvar1.R
\name{predict.test.cartgv}
\alias{predict.test.cartgv}
\title{predict.test.cartgv}
\usage{
\method{predict}{test.cartgv}(new, tree, carts, coups)
}
\arguments{
\item{new}{new tree}

\item{tree}{old tree}

\item{carts}{classification and regression trees}

\item{coups}{coups}
}
\value{
matrix for every new item (sorted) :
\itemize{
\item hat.Y : predicted class
\item Y : la classe d'appartenance
\item noeuds : le numero du noeud contenant l'observation
\item score : le score le l'observation
}
}
\description{
Version of predict.gartgv using tapply instead of for loop
}
