% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_CARTGV.R
\name{predict_cartgv}
\alias{predict_cartgv}
\title{predict_cartgv}
\usage{
predict_cartgv(new, tree, carts, coups)
}
\arguments{
\item{new}{a new data frame containing the same variables that "\code{data}".}

\item{tree}{the data frame "\code{tree}" returned by the function "\code{cartgv}".}

\item{carts}{the list "\code{carts}" returned by the function "\code{cartgv}".}

\item{coups}{the list "\code{table_coupures}" returned by the function "\code{cartgv}".}
}
\value{
a matrix with "\code{nrows(new)}" rows (the i-th row is provided prediction information about the -ith observation in "\code{new}")
and the 4 followingcolumns,
- hat.Y: the predicted label ("0" or "1"),
- Y: the true label ("0" or "1"),
- noeuds: name of the node in the CARTGV tree,
- score:  conditionnal probability that the observation has "Y=1",
- i_noeuds: name of the node in the splitting tree.
}
\description{
Prediction of test data from a fitted CARTGV tree.
}
\details{
The function called the function "\code{pred_cart}".
}
