% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfgv_dataset.R
\docType{data}
\name{rfgv_dataset}
\alias{rfgv_dataset}
\title{Data to test function rfgv}
\format{A data frame with 1000 rows and 27 variables:}
\usage{
rfgv_dataset
}
\description{
A dataset generated to check that fucntion rfgv works.
}
\keyword{datasets}
