% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Species_to_predict}
\alias{Species_to_predict}
\title{Species to predict}
\format{A data frame with 50 rows and 4 variables:
\describe{
\item{Sepal.Length}{ Sepal length}
\item{Sepal.Width}{  Sepal width}
\item{Petal.Length}{ Petal length}
\item{Petal.Width}{ Petal width}
}}
\usage{
Species_to_predict
}
\description{
A dataset from Lise Vaudor blog post about random forest
It is link with iris dataset
}
\keyword{datasets}
