% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions_CARTGVRF_sampvar1.R
\name{split.cartgv}
\alias{split.cartgv}
\title{split.cartgv()}
\usage{
\method{split}{cartgv}(node, group, igroups, label, maxdepth = 2,
  penalty = "No", sampvar = "FALSE",
  mtry_var = sapply(as.numeric(table(group[!is.na(group)])), function(x)
  floor(sqrt(x))))
}
\arguments{
\item{node}{}

\item{group}{}

\item{igroups}{}

\item{label}{}

\item{maxdepth}{}

\item{penalty}{}

\item{sampvar}{}

\item{mtry_var}{}
}
\description{
split.cartgv()
}
