% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions_CARTGVRF_sampvar1.R
\name{split_cartgv}
\alias{split_cartgv}
\title{split_cartgv()}
\usage{
split_cartgv(node, group, igroups, label, maxdepth = 2, penalty = "No",
  sampvar = "FALSE",
  mtry_var = sapply(as.numeric(table(group[!is.na(group)])), function(x)
  floor(sqrt(x))))
}
\arguments{
\item{node}{node}

\item{group}{group}

\item{igroups}{group id}

\item{label}{label}

\item{maxdepth}{maxdepth}

\item{penalty}{penalty}

\item{sampvar}{sample variable}

\item{mtry_var}{mtry variable}
}
\value{
list
}
\description{
split_cartgv()
}
