% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_calculation_surrogate_splits.R
\name{surrogate_split}
\alias{surrogate_split}
\title{surrogate_split}
\usage{
surrogate_split(Ystar, node, group, igroup, penalty = "No", label,
  maxdepth)
}
\arguments{
\item{Ystar}{Ystar}

\item{node}{node}

\item{group}{group}

\item{igroup}{igroup}

\item{penalty}{penalty}

\item{label}{label}

\item{maxdepth}{maxdepth}
}
\value{
a list
}
\description{
surrogate_split
}
