% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algotlda-package.R
\docType{package}
\name{tplda-package}
\alias{tplda}
\alias{tplda-package}
\title{tplda: Tree Penalized Linear Discriminant Analysis}
\description{
Classification tree algorithm for grouped variables
}
\keyword{internal}
