% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPhoneme_test.R
\docType{data}
\name{tree.cartgv}
\alias{tree.cartgv}
\title{Dataframe contains in data_phoneme}
\format{data frame}
\usage{
tree.cartgv
}
\description{
Dataframe contains in data_phoneme
}
\keyword{datasets}
