% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_performance.R
\name{xtab_function}
\alias{xtab_function}
\title{xtab function}
\usage{
xtab_function(predY, Y)
}
\arguments{
\item{predY}{A vector containing the predictions}

\item{Y}{A vector containing the true class labels. Must have the same dimensions as ’predY’}
}
\value{
xtab A list with elements
xtab A list with elements: , , ,
\itemize{
\item table : the results of table on data and reference
\item positive : the positive result level
\item overall : a numeric vector with overall accuracy and Kappa statistic values
\item byClass : the sensitivity, specificity, positive predictive value, negative predictive value, precision, recall, F1,
prevalence, detection rate, detection prevalence and balanced accuracy for each class. For two class systems,
this is calculated once using the positive argument
}
}
\description{
Calculates a cross-tabulation of observed and predicted classes with associated statistics
}
\details{
Compared to the function caret::confusionMatrix, this function deals with the case where a model assigns all observation to the same class (0 or 1)
}
