% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Cost_Complexity_Pruning.R
\name{best_node_cost}
\alias{best_node_cost}
\title{best_node_cost}
\usage{
best_node_cost(node, tree)
}
\arguments{
\item{node}{the number of the node in the tree \code{tree}}

\item{tree}{a tree. A data frame returned by the function \code{CARTGV} and which summarizes the resulted CARTGV tree.}
}
\value{
a numeric that gives the minimum change in the subtree stremming from \code{node}
}
\description{
function that gives the minimum change in tree cost ==> gives alpha = min g(t), for all t in T
}
