% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_RFGV.R
\name{bsamples}
\alias{bsamples}
\title{bsamples}
\usage{
bsamples(ntree, data, sampsize, replace)
}
\arguments{
\item{ntree}{number of boostrap samples (i.e. the number of trees in the RFGV forest)}

\item{data}{the learning data set. it must be a data frame}

\item{sampsize}{an interger indicating the size of the boostrap samples}

\item{replace}{a boolean indicating if sampling of cases is done with or without replacement?}
}
\value{
a list with elements
- bsamples:  a matrix which with '\code{sampsize}' lines and '\code{B}' columns which contains the indices of the observations
belonging to each boostrap sample
- oobsamples: a list that contains the indices of the observations belonging to each out-of-bag sample
}
\description{
Generates the bootstrap samples used to built a RFGV forest
}
