% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions_RPART_sup.R
\name{calcul_cart}
\alias{calcul_cart}
\title{Reconstruction  des arbres CART}
\usage{
calcul_cart(cart, data)
}
\arguments{
\item{cart}{objet rpart}

\item{data}{data.frame ayant servi à la construction de l'objet cart}
}
\value{
liste  :
\itemize{
\item tableau : un data.frame qui donne toutes les infos sur l'arbre cart
("node_name_cart"= numero du noeud,"parent"=ancêtre,"depth"=profondeur du noeud,
"var"=variable utilisée pour couper le noeud,"threshold"=seuil de la coupure,
"n"=effectif du noeud,"pred"=label du neoud,"n_noncase"= nombre d'observation "Y=0" dans le noeuds
,"n_case"=nombre d'individus "Y=1" dans le noeud,"prob"= P\link{Y=1|N},
"sens"=sens pour la règle de coupure divisant son noeud parent:
le sous ensembles des observations verifiant x<threshold est envoyé à gauche si "inf" à droite sinon. )
}
}
\description{
La fonction construit un data.frame qui permet de donner toutes les infos necessaire à
la reconstruction de l'arbre cart construit par RPART.
Attention cette fonction n'est valide que si rpart est parametré
avec control=rpart.control(maxsurrogate=0, maxcompte=0)
}
