% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions_CARTGVRF_sampvar1.R
\name{cartgv}
\alias{cartgv}
\title{cartgv}
\usage{
cartgv(data, group, crit = 1, case_min = 1, maxdepth = 2,
  RF = FALSE, penalty = "No", p = ifelse(!is.null(mtry_group),
  mtry_group, sqrt(length(unique(group[!is.na(group)])))),
  IMPORTANCE = ifelse(RF == TRUE, FALSE, TRUE), sampvar = ifelse(RF ==
  TRUE, TRUE, FALSE),
  mtry_var = sapply(as.numeric(table(group[!is.na(group)])), function(x)
  floor(sqrt(x))))
}
\arguments{
\item{data}{the trainning dataset}

\item{group}{a vector with the group number of each variable (WARNING : if there are P goups, then the vector group must contains all the numbers from 1 to P )}

\item{crit}{1=Gini,2=Entropie,3=Misclassification error}

\item{case_min}{minimum number of cases/non cases in a node which allows to split the node}

\item{maxdepth}{the max depth for a split-tree}

\item{RF}{Random Forest}

\item{penalty}{"No","Size","Root.size" or "Log"}

\item{p}{p parameter}

\item{IMPORTANCE}{importance}

\item{sampvar}{a boolean indicating if within each tree-split, a subset of variables is drawn for each group}

\item{mtry_var}{usefull only if sampvar=TRUE. It indicates the number of drawn variables}
}
\value{
\itemize{
\item tree : data.frame "resumant" l'arbre construit
\item carts : liste des arbres carts utilises pour elaborer l'arbre
\item splits : liste donnant pour chaque noeud la valeur de la fonction Split_Tree_CART()
\item pop : liste contenant pour chaque neoud le nom des individus (rownames)
\item tables_coupures : liste contenant pour chaque coupure un data.frame permettant de reconstruire l'arbre CART de coupure
\item groups_selec   : liste contenant pour chaque coupure les groupes candidats
\item importance     : liste donnant pour chaque groupe la décroissance d'impureté resultant de la meilleur coupure opérée sur le groupe
\item agreement1      : liste donnant pour chaque coupure et chaque groupe candidat la proba d'accord (Indice de Rand) entre la coupure choisie et la meilleure coupure pour le groupe
\item cimportance1     : liste donnant pour chaque coupure et chaque groupe l'importance "corrigée" par la proba d'accord
}
}
\description{
Classification And Regression Trees for Grouped Variables
}
\details{
Function which takes as inputs a integer "B"  and a sample of data "data" . The function
draws randomly B bootstrap samples of size N (the size of the sample of data) and returns
a matrix which with N lines and B columns which contains the indices of the observations
belonging each boostrap sample.
}
