% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions_CARTGVRF_sampvar2.R
\name{cartgv.rf}
\alias{cartgv.rf}
\title{cartgv.rf}
\usage{
cartgv.rf(data, group, crit = 1, case_min = 1, maxdepth = 2,
  p = floor(sqrt(length(unique(group[!is.na(group)])))),
  penalty = "No",
  mtry_var = sapply(as.numeric(table(group[!is.na(group)])), function(x)
  floor(sqrt(x))))
}
\arguments{
\item{data}{data}

\item{group}{group}

\item{crit}{criteria}

\item{case_min}{miniaml case}

\item{maxdepth}{maximum depth}

\item{p}{parameter p}

\item{penalty}{penalty}

\item{mtry_var}{mtry variable}
}
\value{
list
}
\description{
Classification Random Tree for Grouped Variables using Random Forest.
}
\details{
A utiliser quand sampvar=TRUE and sampvar_type="2 ==> permet de construire un arbre
cartgv tel que dans chaque arbre de
coupure un sous-ensemble de variables est tiré au hasard avant chaque coupure
}
