% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions_CARTGVRF_sampvar2.R
\name{cartgv.rf}
\alias{cartgv.rf}
\title{Title}
\usage{
cartgv.rf(data, group, crit = 1, case_min = 1, maxdepth = 2,
  p = floor(sqrt(length(unique(group[!is.na(group)])))),
  penalty = "No",
  mtry_var = sapply(as.numeric(table(group[!is.na(group)])), function(x)
  floor(sqrt(x))))
}
\arguments{
\item{data}{}

\item{group}{}

\item{crit}{}

\item{case_min}{}

\item{maxdepth}{}

\item{p}{}

\item{penalty}{}

\item{mtry_var}{}
}
\description{
Title
}
