% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Cost_Complexity_Pruning.R
\name{cost_error}
\alias{cost_error}
\title{cost_error}
\usage{
cost_error(node, tree, alpha)
}
\arguments{
\item{node}{the number of the node in the tree \code{tree}}

\item{tree}{a tree. A data frame returned by the function \code{CARTGV} and which summarizes the resulted CARTGV tree.}

\item{alpha}{a positive real. It is the value of the complexity paramter in the cost-complexity criterion.}
}
\value{
a numeric indicating the misclassification error of the subtree stremming from node penalized by alpha.
}
\description{
function that calculate the misclassification error of the subtree stremming from \code{node} penalized by alpha
}
