% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Cost_Complexity_Pruning.R
\name{descendant}
\alias{descendant}
\title{descendant}
\usage{
descendant(node, tree)
}
\arguments{
\item{node}{the number of the node in the tree \code{tree}.}

\item{tree}{a tree. A data frame returned by the function \code{CARTGV} and which summarizes the resulted CARTGV tree.}
}
\value{
descendants: a vector with the number of the descendants of \code{node} in \code{tree}.
}
\description{
function that find the direct descendants of a node
}
