% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions_RF_evo_Modif_pierre.R
\name{entropy}
\alias{entropy}
\title{Impurity functions}
\usage{
entropy(p)
}
\arguments{
\item{p}{probability}
}
\value{
entropy
}
\description{
works for the case when y has only 0 and 1 categorie...
}
