% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Cost_Complexity_Pruning.R
\name{extract_subtrees}
\alias{extract_subtrees}
\title{extratct_subtrees}
\usage{
extract_subtrees(tree, cp)
}
\arguments{
\item{tree}{}

\item{cp}{}
}
\description{
function that extract the sequence of nested trees
}
