% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Cost_Complexity_Pruning.R
\name{get_nodes}
\alias{get_nodes}
\title{get_nodes}
\usage{
get_nodes(node, tree)
}
\arguments{
\item{node}{}

\item{tree}{}
}
\description{
function that gives the nodes of the subtree stremming from node
}
\details{
WARNINGS: this function was modified on the 4th december 2017 to be adapted to non-binary trees.
}
