% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{group}
\alias{group}
\title{Data to test function rfgv}
\format{An object of class \code{numeric} of length 26.}
\usage{
group

group
}
\description{
A dataset generated to check that fucntion rfgv works.
Please load with rfgv_dataset
}
\details{
26 values
}
\keyword{datasets}
