% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_RFGV.R
\name{group.selection}
\alias{group.selection}
\title{group.selection}
\usage{
group.selection(group, mtry = sqrt(unique(group[!is.na(group)])))
}
\arguments{
\item{group}{a vector indicating the group label of each variable}

\item{mtry}{a number indicating the number of selected group}
}
\value{
groups a vector indicating the label of the selected groups
}
\description{
Selects randomly '\code{mtry}' groups
}
