% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions_CARTGVRF_sampvar1.R
\name{grpimpgini}
\alias{grpimpgini}
\title{grpimpgini}
\usage{
grpimpgini(num_group, groupselec, tree)
}
\arguments{
\item{num_group}{group number}

\item{groupselec}{selection group}

\item{tree}{tree}
}
\value{
DecreaseImpurity
}
\description{
grpimpgini
}
