% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions_CARTGVRF_sampvar1.R
\name{grpimpperm}
\alias{grpimpperm}
\title{grpimpperm}
\usage{
grpimpperm(num_group, data, oobsamples, group, tree, impurityacc)
}
\arguments{
\item{num_group}{group number}

\item{data}{input data}

\item{oobsamples}{out of bag samples}

\item{group}{group}

\item{tree}{tree}

\item{impurityacc}{impurity cumsum}
}
\description{
grpimpperm
}
