% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions_CARTGVRF_sampvar2.R
\name{grpimpperm.rf}
\alias{grpimpperm.rf}
\title{grpimpperm.rf}
\usage{
grpimpperm.rf(num_group, data, oobsamples, group, tree, impurityacc)
}
\arguments{
\item{num_group}{group number}

\item{data}{input data}

\item{oobsamples}{out of bag samples}

\item{group}{group id}

\item{tree}{tree data}

\item{impurityacc}{impurity cumsum}
}
\value{
DecreaseAcc
}
\description{
grpimpperm.rf
}
