% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions_RPART_sup.R
\name{impurete_rpart}
\alias{impurete_rpart}
\title{Elagage}
\usage{
impurete_rpart(validation, tree_seq)
}
\arguments{
\item{validation}{}

\item{tree_seq}{}
}
\value{
\itemize{
\item impurete : une matrice contenant les differentes valeurs d'impuret'e pour chaque sous-arbres
\item pred = liste des predictions pour chaque sous arbres
\item summary_noeuds : liste contenant pour chaque sous-arbre des infos sur ses noeuds:
\itemize{
\item nom_noeuds = noms du noeuds
\item N = nb obs dans le neoud
\item \code{N[Y=1]} = nb d'obs dans le noeuds appartenant a' la class "Y=1"
\item \code{P[Y=1]} = proba pour une obs du noeuds d'appartenanir  a' la class "Y=1"
\item \code{P[Y=0]} = proba pour une obs du noeuds d'appartenanir  a' la class "Y=0"
\item \code{P[hat.Y!=Y]} = tx de mal class'es dans le noeud
}
}
}
\description{
calcul de l'impurete
}
\details{
a partir d'un echantillon independant et d'une sequence d'arbres emboites
La fonction prend en entree un échantillon test et une sequence d'arbres emboites
Pour chaque arbre, on predit la classe de chaque observation de l'ensemble test;
puis a partir de ces resultats, on calcul l'impurete de chaque arbre (gini, l'entropie et
le taux de mal-classes sont calcules)
}
