% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_RPART_tools.R
\name{impurete_rpart}
\alias{impurete_rpart}
\title{impurete_rpart}
\usage{
impurete_rpart(validation, tree_seq)
}
\arguments{
\item{validation}{an new data set. It must be a data frame containing the same variables that those contained in the learning data set used to built the rpart object.}

\item{tree_seq}{a sequence of subtrees. It must be a list where each element is an object returned by the function 'rpart::snip.rpart'}
}
\value{
a list with elements
\itemize{
\item impurete:  a matrix containing the impurity values (respectively Gini, Entropy and Misclassification rate) of each subtrees evaluated on the data set '\code{validation}'.
\item pred:      a list containing the prediction vector of each subtree
\item summary_noeuds:  a list providing informations about each subtree:
\itemize{
\item nom_noeuds: number of the node
\item N: number of observations in the node
\item \code{N[Y=1]}: number of observation with "\code{Y=1}" in the node
\item \code{P[Y=1]}: empirical probability that in the node an observation belongs to the label "\code{Y=1}"
\item \code{P[Y=0]}: empirical probability that in the node an observation belongs to the label "\code{Y=0}"
\item \code{P[hat.Y!=Y]}: misclassification rate in the node
}
}
}
\description{
Compute the impurity of a sequence of trees which are based on a rpart object.
}
\details{
a partir d'un echantillon independant et d'une sequence d'arbres emboites
La fonction prend en entree un échantillon test et une sequence d'arbres emboites
Pour chaque arbre, on predit la classe de chaque observation de l'ensemble test;
puis a partir de ces resultats, on calcul l'impurete de chaque arbre (gini, l'entropie et
le taux de mal-classes sont calcules)
}
