% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions_CARTGVRF_sampvar2.R
\name{impurity.cartgv.rf}
\alias{impurity.cartgv.rf}
\title{Title}
\usage{
impurity.cartgv.rf(validation, tree_seq, tree)
}
\arguments{
\item{validation}{}

\item{tree_seq}{}

\item{tree}{}
}
\description{
Title
}
