% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Cost_Complexity_Pruning.R
\name{node_cost}
\alias{node_cost}
\title{node_cost}
\usage{
node_cost(node, tree)
}
\arguments{
\item{node}{}

\item{tree}{}
}
\description{
function that gives the change in tree cost (difference between the node error and the local node error)  ==> gives alpha (function g(t) in the Breiman's book)
}
