% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Cost_Complexity_Pruning.R
\name{node_error}
\alias{node_error}
\title{node_error}
\usage{
node_error(node, tree)
}
\arguments{
\item{node}{}

\item{tree}{}
}
\description{
Function that calculate the misclassification error of the subtree stremming from node
}
