% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Cost_Complexity_Pruning.R
\name{node_prob}
\alias{node_prob}
\title{node_prob}
\usage{
node_prob(node, tree)
}
\arguments{
\item{node}{}

\item{tree}{}
}
\description{
function that calculate the probability of being at the node
}
