% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions_performance.R
\name{opt.cut}
\alias{opt.cut}
\title{perf}
\usage{
opt.cut(perf)
}
\arguments{
\item{perf}{}
}
\description{
Optimal cut
}
\details{
Trouver le seuil qui permet le meilleur compromis sensibilite-specificite
}
