% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions_CARTGVRF_sampvar1.R
\name{perm}
\alias{perm}
\title{perm}
\usage{
perm(oobsamples, data, num_group, group)
}
\arguments{
\item{oobsamples}{out of bag samples}

\item{data}{data}

\item{num_group}{group number}

\item{group}{group}
}
\value{
data_perm
}
\description{
perm
}
