% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions_CARTGVRF_sampvar2.R
\name{predict.cartgv.rf}
\alias{predict.cartgv.rf}
\title{Title}
\usage{
\method{predict}{cartgv.rf}(new, tree, tree_split)
}
\arguments{
\item{new}{}

\item{tree}{}

\item{tree_split}{}
}
\description{
Title
}
