% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions_CARTGVRF_sampvar1.R
\name{predict.test.cartgv}
\alias{predict.test.cartgv}
\title{Title}
\usage{
\method{predict}{test.cartgv}(new, tree, carts, coups)
}
\arguments{
\item{new}{}

\item{tree}{}

\item{carts}{}

\item{coups}{}
}
\value{
la fonction renvoie une matrice donnant pour chaque
individu de new (dans l'ordre) :
\itemize{
\item hat.Y :la classe predite
\item Y : la classe d'appartenance
\item noeuds : le numero du noeud contenant l'observation
\item score : le score le l'observation
}
}
\description{
Title
}
