% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions_CARTGVRF_sampvar1.R
\name{predict_cartgv}
\alias{predict_cartgv}
\title{predict_cartgv}
\usage{
predict_cartgv(new, tree, carts, coups)
}
\arguments{
\item{new}{new tree}

\item{tree}{old tree}

\item{carts}{calssification and regression tree}

\item{coups}{cuts}
}
\value{
la fonction renvoie une matrice donnant pour chaque individu de new (dans l'ordre) :
\itemize{
\item hat.Y :la classe predite
\item Y : la classe d'appartenance
\item noeuds : le numero du noeud contenant l'observation
\item score : le score le l'observation
}
}
\description{
Predict classification and regression tree for grouped variables
}
\details{
la fonction fait appel a la fonction pred_cart()
IMPORTANT : i_noeuds (indice du noeud dans l'arbre de coupure, cad indice donne par rpart)
et noeuds (indice du noeud dans l'arbre cartgv) utilisees comme cles primaires pour
identifier de manière unique un noeuds
}
