% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_variant_CARTGV_for_RFGV.R
\name{predict_cartgv.rf}
\alias{predict_cartgv.rf}
\title{predict_cartgv.rf}
\usage{
predict_cartgv.rf(new, tree, tree_split)
}
\arguments{
\item{new}{a new data frame containing the same variables that "\code{data}".}

\item{tree}{the data frame "\code{tree}" returned by the function "\code{cartgv.rf}".}

\item{tree_split}{the object "\code{tree_split}" returned by the function "\code{cartgv.rf}".}
}
\value{
a matrix with "\code{nrows(new)}" rows (the i-th row is provided prediction information about the -ith observation in "\code{new}")
and the 4 followingcolumns,
- Y: the true label ("0" or "1"),
- hat.Y: the predicted label ("0" or "1"),
- noeuds: name of the node in the modified CARTGV tree,
- score:  conditionnal probability that the observation has "Y=1",
- i_noeuds: name of the node in the splitting tree.
}
\description{
Prediction of test data from a fitted modified CARTGV tree.
}
\details{
The function called the function "\code{predict_cartgv}".
}
