% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfgv_dataset.R
\docType{data}
\name{rfgv_dataset}
\alias{rfgv_dataset}
\title{Data to test function rfgv}
\format{A data frame with 1000 rows and 27 variables:
\describe{
\item{\code{Type}}{train, test, validation}
\item{\code{Y}}{0 or 1}
\item{\code{V3_G1}}{ some values }
\item{\code{V4_G1}}{ some values }
\item{\code{V5_G1}}{ some values }
\item{\code{V6_G1}}{ some values }
\item{\code{V7_G1}}{ some values }
\item{\code{V8_G2}}{ some values }
\item{\code{V9_G2}}{ some values }
\item{\code{V10_G2}}{ some values }
\item{\code{V11_G2}}{ some values }
\item{\code{V12_G2}}{ some values }
\item{\code{V13_G3}}{ some values }
\item{\code{V14_G3}}{ some values }
\item{\code{V15_G3}}{ some values }
\item{\code{V16_G3}}{ some values }
\item{\code{V17_G3}}{ some values }
\item{\code{V18_G4}}{ some values }
\item{\code{V19_G4}}{ some values }
\item{\code{V20_G4}}{ some values }
\item{\code{V21_G4}}{ some values }
\item{\code{V22_G4}}{ some values }
\item{\code{V23_G5}}{ some values }
\item{\code{V24_G5}}{ some values }
\item{\code{V25_G5}}{ some values }
\item{\code{V26_G5}}{ some values }
\item{\code{V27_G5}}{ some values }
}}
\usage{
rfgv_dataset
}
\description{
A dataset generated to check that fucntion rfgv works.
}
\keyword{datasets}
