% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_CARTGV.R
\name{split_cartgv}
\alias{split_cartgv}
\title{split_cartgv()}
\usage{
split_cartgv(node, group, label, maxdepth = 2, penalty = "No")
}
\arguments{
\item{node}{a data frame containing the observations in the node. The first column is the response vector named "Y" and with the lable
"0" and "1". The p-1 others variables are continuous or categroical variable must be coded as a set of dummy variables.}

\item{group}{a vector with the group number of each variable.
(WARNING : if there are "\code{p}" goups, the groups must be numbers from "\code{1}" to "\code{p}" in increasing order. The group label
of the response variable is missing (i.e. NA)).}

\item{label}{an integer indicating the label of the node (the majority class)}

\item{maxdepth}{an integer indicating the maximal depth for a split-tree. The default value is 2.}

\item{penalty}{a boolean indicating if the decrease in node impurity must take account of the group size. Four penalty are available: "No"
,"Size","Root.size" or "Log".}
}
\value{
a list with elements
- Gain_Gini: a vector containing the reduction of Gini in the node from splitting on each group,
- Gain_Ent: a vector containing the reduction of Entropy in the node from splitting on each group,
- Gain_Mis: a vector containing the reduction of the number of miclassified observations in the node from splitting on each group,
- carts: a list containing for each group the CART object which summarizes the splitting tree,
- pred: a matrix with "\code{nrows(node)}" lines and "\code{length(unique(group))}" columns containing for each group the prediction,
resulting from the splitting tree.
}
\description{
Calculate the best split of a node for each group of input variables when building a CARTGV tree.
}
