% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_calculation_surrogate_splits.R
\name{surrogate_split}
\alias{surrogate_split}
\title{surrogate_split}
\usage{
surrogate_split(Ystar, node, group, igroup, penalty = "No", label,
  maxdepth)
}
\arguments{
\item{Ystar}{}

\item{node}{}

\item{group}{}

\item{igroup}{}

\item{penalty}{}

\item{label}{}

\item{maxdepth}{}
}
\description{
surrogate_split
}
