% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions_performance.R
\name{xtab_function}
\alias{xtab_function}
\title{xtab function}
\usage{
xtab_function(predY, Y)
}
\arguments{
\item{predY}{}

\item{Y}{}
}
\description{
Matrice de confusion + performances predictives
}
\details{
(gere quand le modele predit tout en 1 ou 0)
}
